Fix the cpu core count computation
The original core count computation used grep to find a certain row from
the output of lscpu. Now the output of lscpu has changed, so the regex
used by grep found two entries. I'm not sure how to fix the regex in a
way that works in both Swedish and English. So instead I switched to
finding the core count from the file /proc/cpuinfo. The regex used to
find the right row in that file is looking for the word 'processor' and
should therefore work with the same regex in both Swedish and English.